restructure component and styles folders
