+++ src/components/index.scss
@@ -1,0 +1,5 @@
+@import "./header/header";
+@import "./login/login";
+@import "./playlistComponents/playlistPreview/playlistPreview";
+@import "./playlistComponents/seedArtists/seedArtists";
+@import "./playlistComponents/playlist";
