+++ src/components/login/login.scss
@@ -1,0 +1,17 @@
+@import "@sass/mixins.scss";
+
+/*====================
+Login Screen
+=====================*/
+
+.login-header {
+    text-align: center;
+    color: #fefefe;
+    font-size: $font-md;
+    margin-bottom: $gutter-md;
+}
+
+.login-container {
+    margin-top: 15rem;
+    text-align: center;
+}
