+++ src/components/playlistComponents/playlist.scss
@@ -1,0 +1,33 @@
+@import "@sass/mixins";
+
+/*===================
+Playlist Pages
+====================*/
+
+.form-container {
+    margin: 10rem auto 0 auto;
+    text-align: center;
+    max-width: 345px;
+}
+
+.form-container h3 {
+  font-size: $font-md;
+  margin-bottom: $gutter-md;
+}
+
+.form-container input {
+  padding: $input-padding;
+  min-width: 250px;
+  max-width: 400px;
+  width: 70%;
+  display: block;
+  margin: 0 auto $gutter-sm;
+  box-shadow: inset -2px -2px 4px rgb(0 0 0 / 25%), inset 3px 0 4px rgb(0 0 0 / 24%);
+  border: 1px solid #111;
+}
+
+@media screen and (min-width: 599px){
+  .form-container {
+    max-width: 100%;
+  }
+}
