+++ src/components/playlistComponents/playlistPreview/playlistPreview.scss
@@ -1,0 +1,68 @@
+@import "@sass/mixins.scss";
+
+/*========================
+Playlst Preview
+=========================*/
+
+.song-info {
+    display: block;
+    padding-left: 2rem;
+    text-align: left;
+    width: 100%;
+    & h4 {
+      color: $primary-green;
+      font-size: $font-lg;
+    }
+    & p {
+      font-size: $font-sm;
+      color: rgb(254 254 254 / 85%);
+    }
+}
+
+.album-cover {
+  width: 185px;
+  margin-bottom: 0;
+}
+
+@media screen and (max-width: 599px){
+  .album-cover {
+    width: 100px;
+  }
+
+  .song-info h4 {
+    font-size: $font-md;
+  }
+}
+
+.song {
+  display: flex;
+  align-items: center;
+  margin: 3rem 0;
+  text-align: left;
+}
+
+.button-holder {
+    display: flex;
+    height: 48px;
+    align-self: center;
+    align-items: center;
+    & span {
+      color: rgb(254 254 254 / 75%);
+      margin: 0 7px;
+      font-size: $font-sm;
+    }
+    & button {
+      background: none;
+      border: none;
+    }
+    & .remove {
+      color: red;
+    }
+    & .add {
+      color: $primary-green;
+    }
+}
+
+.inactive {
+  opacity: 0.5;
+}
