+++ src/components/playlistComponents/seedArtists/seedArtists.scss
@@ -1,0 +1,59 @@
+@import "@sass/mixins.scss";
+
+/*========================
+Select Artists Page
+=========================*/
+
+.artist-img {
+   border-radius: 50%;
+   margin-right: 1rem;
+   width: 125px;
+   height: 125px;
+   vertical-align: middle;
+   display: inline-block;
+}
+
+.error-mesage {
+    padding: 0.8rem 2rem;
+    background: #b93737;
+    max-width: 575px;
+    margin: 0 auto 2rem;
+    opacity: 1;
+    border-radius: 30px;
+    transition: all 0.2s 0.2s linear;
+}
+
+.remove-button {
+  float: right;
+  margin-left: 15px;
+}
+
+.search-dropdown {
+  margin-top: $gutter-lg;
+  & li {
+    list-style: none;
+    text-align: left;
+    color: #fefefe;
+    margin-bottom: $gutter-sm;
+  }
+}
+
+.artist {
+    padding: 1rem 2rem;
+    box-shadow: 2px 2px 2px 4px #20202010;
+    &.active {
+      box-shadow: 1px 1px 2px 4px #5dbb60;
+    }
+}
+
+
+
+.artists  {
+  max-width: 550px;
+  margin: 0 auto;
+}
+
+.selected-artists {
+  list-style: none;
+  color: #fefefe;
+}
