+++ src/components/userProfile/userProfile.scss
@@ -1,0 +1,64 @@
+@import "@sass/mixins";
+
+/*====================
+User Profile
+=====================*/
+
+.hidden {
+  display: none;
+}
+
+.active {
+  display: block;
+}
+
+.search-container {
+  margin: 0 auto;
+  max-width: 960;
+  padding: 1.45rem 1.0875rem;
+  text-align: center;
+}
+
+.btn.btn-success {
+  padding: .375rem 2.75rem;
+  font-size: 1rem;
+  line-height: 1.5;
+  border-radius: 1.25rem;
+}
+
+.search-input.form-control {
+  border-radius: 30px;
+  font-size: 1.2rem;
+  margin-bottom: 2rem;
+}
+
+.user-info {
+    text-align: left;
+    position: absolute;
+    left: 5%;
+    top: 29px;
+    & h2 {
+      font-weight: 400;
+      font-family: 'Open Sans', sans-serif;
+      font-size: 0.875rem;
+      margin-bottom: 0;
+      display: inline-block;
+      margin-left: $gutter-sm;
+    }
+    & span {
+      padding: 0.45rem 1rem 0.75rem;
+      width: 50px;
+      height: 50px;
+      border-radius: 50%;
+      background: #5dbb60;
+      font-weight: 600;
+      display: inline-block;
+      font-size: 1.2rem;
+    }
+}
+
+@media screen and (max-width: 599px){
+  .user-info .hidden-sm{
+    display: none;
+  }
+}
