--- src/data/test.js
@@ -1,26 +1,0 @@
-export const results = {
-  city: "Berlin",
-  artists: [
-    "Daft Punk",
-    "Ross From Friends",
-    "Mallgrab",
-    "Flume",
-    "Rufus Du Sol",
-    "Lane 8",
-    "Tame Impala",
-    "Arctic Monkeys",
-    "Jaime XX"
-  ],
-  genres: [
-    "House",
-    "Techno",
-    "Alternative Rock",
-    "Electronic"
-  ],
-  characteristics: [
-    "dance",
-    "upbeat",
-    "high tempo",
-    "experimental"
-  ]
-}
