+++ src/pages/playlist.js
@@ -4,9 +4,9 @@ import 'bootstrap/dist/css/bootstrap.min.css'
 
 import Container from "react-bootstrap/Container"
 import Row from "react-bootstrap/Row"
-import Layout from "../components/layout"
-import PlaylistContainer from "../components/PlaylistContainer"
-import SEO from "../components/seo"
+import Layout from "@components/gatsbyComponents/layout/Layout"
+import Playlist from "@components/PlaylistComponents/Playlist"
+import SEO from "@components/gatsbyComponents/seo/Seo"
 
 const PlaylistPage = () => (
 
@@ -15,7 +15,7 @@ const PlaylistPage = () => (
     <Container>
       <Row className="justify-content-center">
         <Router>
-          <PlaylistContainer path=":code"/>
+          <Playlist path=":code"/>
         </Router>
       </Row>
     </Container>
