+++ src/pages/using-typescript.tsx
@@ -2,8 +2,8 @@
 import React from "react"
 import { PageProps, Link, graphql } from "gatsby"
 
-import Layout from "../components/layout"
-import SEO from "../components/seo"
+import Layout from "@components/gatsbyComponents/layout/Layout"
+import SEO from "@components/gatsbyComponents/seo/Seo"
 
 type DataProps = {
   site: {
