--- src/styles/custom.scss
@@ -1,334 +1,0 @@
-/*customizations to react-bootstrap styles*/
-@import "~bootstrap/scss/bootstrap";
-
-/*================
-Default Settings
-==================*/
-
-* {
-  font-size: 16px;
-}
-
-#gatsby-focus-wrapper {
-  overflow-x: hidden;
-}
-
-body {
-  background: #121212!important;
-}
-
-p, h2, h3 {
-  color: #efefef;
-}
-
-
-$primary-green: #5DBB60;
-$dark-green: darken($primary-green, 20%);
-$primary-black: #222;
-$button-padding: 0.25rem 2rem 0.5rem 2rem;
-$input-padding: 0.275rem 1rem;
-$gutter-sm: 1rem;
-$gutter-md: 1.5rem;
-$gutter-lg: 2rem;
-$font-lg: 1.4rem;
-$font-md: 1.2rem;
-$font-sm: 1rem;
-
-
-/*==================
-Header section
-===================*/
-
-.header {
-    padding: 1.2rem;
-    text-align: center;
-    width: auto;
-    justify-content: center;
-    & h1 {
-      font-size: $font-lg;
-      font-weight: 700;
-      margin-bottom: 0;
-      color: $primary-green;
-    }
-}
-
-
-/*==================
-Buttons/Typography
-===================*/
-
-.btn.btn-primary {
-  background: $primary-green;
-  border-color: $primary-green;
-  border-radius: 0;
-  padding: $button-padding;
-  & a {
-    color: $primary-black;
-    font-size: $font-sm;
-    font-weight: 700;
-  }
-  &:hover {
-    color: #fff;
-    background-color: $dark-green;
-    border-color: $dark-green;
-  }
-}
-
-h1,h2,h3 {
-  font-family: 'Roboto', sans-serif;
-}
-
-p, a, li {
-  font-family: 'Open Sans', sans-serif;
-}
-
-/*====================
-Login Screen
-=====================*/
-
-.login-header {
-    text-align: center;
-    color: #fefefe;
-    font-size: $font-md;
-    margin-bottom: $gutter-md;
-}
-
-.login-container {
-    margin-top: 15rem;
-    text-align: center;
-}
-
-/*===================
-Playlist Pages
-====================*/
-
-.form-container {
-    margin: 10rem auto 0 auto;
-    text-align: center;
-    max-width: 345px;
-}
-
-.form-container h3 {
-  font-size: $font-md;
-  margin-bottom: $gutter-md;
-}
-
-.form-container input {
-  padding: $input-padding;
-  min-width: 250px;
-  max-width: 400px;
-  width: 70%;
-  display: block;
-  margin: 0 auto $gutter-sm;
-  box-shadow: inset -2px -2px 4px rgb(0 0 0 / 25%), inset 3px 0 4px rgb(0 0 0 / 24%);
-  border: 1px solid #111;
-}
-
-@media screen and (min-width: 599px){
-  .form-container {
-    max-width: 100%;
-  }
-}
-
-/*====================
-User Profile
-=====================*/
-
-.hidden {
-  display: none;
-}
-
-.active {
-  display: block;
-}
-
-.search-container {
-  margin: 0 auto;
-  max-width: 960;
-  padding: 1.45rem 1.0875rem;
-  text-align: center;
-}
-
-.btn.btn-success {
-  padding: .375rem 2.75rem;
-  font-size: 1rem;
-  line-height: 1.5;
-  border-radius: 1.25rem;
-}
-
-.search-input.form-control {
-  border-radius: 30px;
-  font-size: 1.2rem;
-  margin-bottom: 2rem;
-}
-
-.user-info {
-    text-align: left;
-    position: absolute;
-    left: 5%;
-    top: 29px;
-    & h2 {
-      font-weight: 400;
-      font-family: 'Open Sans', sans-serif;
-      font-size: 0.875rem;
-      margin-bottom: 0;
-      display: inline-block;
-      margin-left: $gutter-sm;
-    }
-    & span {
-      padding: 0.45rem 1rem 0.75rem;
-      width: 50px;
-      height: 50px;
-      border-radius: 50%;
-      background: #5dbb60;
-      font-weight: 600;
-      display: inline-block;
-      font-size: 1.2rem;
-    }
-}
-
-@media screen and (max-width: 599px){
-  .user-info .hidden-sm{
-    display: none;
-  }
-}
-
-/*========================
-Select Artists Page
-=========================*/
-
-.artist-img {
-   border-radius: 50%;
-   margin-right: 1rem;
-   width: 125px;
-   height: 125px;
-   vertical-align: middle;
-   display: inline-block;
-}
-
-.error-mesage {
-    padding: 0.8rem 2rem;
-    background: #b93737;
-    max-width: 575px;
-    margin: 0 auto 2rem;
-    opacity: 1;
-    border-radius: 30px;
-    transition: all 0.2s 0.2s linear;
-}
-
-.remove-button {
-  float: right;
-  margin-left: 15px;
-}
-
-.search-dropdown {
-  margin-top: $gutter-lg;
-  & li {
-    list-style: none;
-    text-align: left;
-    color: #fefefe;
-    margin-bottom: $gutter-sm;
-  }
-}
-
-.artist {
-    padding: 1rem 2rem;
-    box-shadow: 2px 2px 2px 4px #20202010;
-    &.active {
-      box-shadow: 1px 1px 2px 4px #5dbb60;
-    }
-}
-
-
-
-.artists  {
-  max-width: 550px;
-  margin: 0 auto;
-}
-
-.selected-artists {
-  list-style: none;
-  color: #fefefe;
-}
-
-/*========================
-Playlst Preview
-=========================*/
-
-.song-info {
-    display: block;
-    padding-left: 2rem;
-    text-align: left;
-    width: 100%;
-    & h4 {
-      color: $primary-green;
-      font-size: $font-lg;
-    }
-    & p {
-      font-size: $font-sm;
-      color: rgb(254 254 254 / 85%);
-    }
-}
-
-.album-cover {
-  width: 185px;
-  margin-bottom: 0;
-}
-
-@media screen and (max-width: 599px){
-  .album-cover {
-    width: 100px;
-  }
-
-  .song-info h4 {
-    font-size: $font-md;
-  }
-}
-
-.song {
-  display: flex;
-  align-items: center;
-  margin: 3rem 0;
-  text-align: left;
-}
-
-.button-holder {
-    display: flex;
-    height: 48px;
-    align-self: center;
-    align-items: center;
-    & span {
-      color: rgb(254 254 254 / 75%);
-      margin: 0 7px;
-      font-size: $font-sm;
-    }
-    & button {
-      background: none;
-      border: none;
-    }
-    & .remove {
-      color: red;
-    }
-    & .add {
-      color: $primary-green;
-    }
-}
-
-.inactive {
-  opacity: 0.5;
-}
-
-
-/*=========================
-Footer
-==========================*/
-
-footer {
-  text-align: center;
-  color: #fefefe;
-  position: static;
-  bottom: 0;
-  margin-top: 15rem;
-  left: 0;
-  width: 100%;
-}
