+++ src/styles/style.scss
@@ -1,1 +1,3 @@
-@import "custom";
+@import "mixins";
+@import "~bootstrap/scss/bootstrap";
+@import "@components/index.scss"
